Move XmlSchemaTestCase assertions to Assert
