+++ src/Extension/Phpunit/Constraint/SchemaAcceptsXml.php
@@ -3,14 +3,13 @@
 /*
  * This file is part of the Symfony CMF package.
  *
- * (c) 2011-2014 Symfony CMF
+ * (c) 2011-2015 Symfony CMF
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-
-namespace Symfony\Cmf\Component\Testing\Unit\Constraint;
+namespace Symfony\Cmf\Component\Testing\Extension\Phpunit\Constraint;
 
 class SchemaAcceptsXml extends \PHPUnit_Framework_Constraint
 {
