+++ tests/Extension/Phpunit/Constraint/SchemaAcceptsXmlTest.php
@@ -3,16 +3,15 @@
 /*
  * This file is part of the Symfony CMF package.
  *
- * (c) 2011-2014 Symfony CMF
+ * (c) 2011-2015 Symfony CMF
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
+namespace Symfony\Cmf\Component\Testing\Tests\Extension\Phpunit\Constraint;
 
-namespace Symfony\Cmf\Component\Testing\Tests\Unit\Constraint;
-
-use Symfony\Cmf\Component\Testing\Unit\Constraint\SchemaAcceptsXml;
+use Symfony\Cmf\Component\Testing\Extension\Phpunit\Constraint\SchemaAcceptsXml;
 
 class SchemaAcceptsXmlTest extends \PHPUnit_Framework_TestCase
 {
